Rename bindingDOM.ts to syntax-tree.ts
