+++ src/parser/index.ts
@@ -2,5 +2,5 @@ import './polyfill'
 export { parse } from './parser'
 export { Reporting } from '../program'
 export { Diagnostic, Severity } from '../diagnostic'
-export { Binding, BindingData, BindingExpression, BindingName, DiagNode, Document, Node, NodeType, ImportNode } from './bindingDOM'
+export { Binding, BindingData, BindingExpression, BindingName, DiagNode, Document, Node, NodeType, ImportNode } from './syntax-tree'
 export { createDocument } from './document-builder'
