+++ src/parser/parser-compiler.ts
@@ -4,7 +4,7 @@ import { Parser } from 'jison'
 import * as fs from 'fs'
 import * as path from 'path'
 import { YY } from './parser'
-import { Node } from './bindingDOM'
+import { Node } from './syntax-tree'
 
 const bnf = fs.readFileSync(path.join(__dirname, '../../src/parser/grammar.jison'), 'utf8')
 const parser = new Parser<YY, Node[]>(bnf /*, { debug: true, type: 'lr' }*/)
